rova: added debug.sf.latch_unsignaled=1
The Android debugging option that is used to reduce the amount of time spent on waiting for the latch to be signaled. This option is used to improve the performance of the Android system by allowing the system to skip waiting for the latch to be signaled and continuing to execute the code.