rova: add debug.sf.disable_backpressure=1
This setting in Android is a debugging option that disables the backpressure mechanism for the SurfaceFlinger layer. This mechanism is used to reduce the amount of work the GPU does when the system is overloaded. By disabling this mechanism, the GPU will be able to do more work and the overall performance of the system may improve.